+++ static/js/font-loader.js
@@ -4,10 +4,3 @@
     wf.async = true;
     wf.src = 'https://ajax.googleapis.com/ajax/libs/webfont/1.6.26/webfont.js';
 })(document);
-
-(function (d) {
-    var fa = d.createElement('link');
-    fa.rel = "stylesheet";
-    fa.href = "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css";
-    d.head.appendChild(fa);
-})(document);